:recycle: Rename package to comply with pmd
